+++ src/eval/bc/vm.c
@@ -423,6 +423,7 @@ static mino_val *bc_protocol_type_disc(mino_state *S, mino_val *v)
     case MINO_TX_REF:     return mino_keyword(S, "ref");
     case MINO_AGENT:      return mino_keyword(S, "agent");
     case MINO_CHAN:       return mino_keyword(S, "chan");
+    case MINO_QUEUE:      return mino_keyword(S, "queue");
     }
     return mino_keyword(S, "unknown");
 }